bugfix: 수익률 계산 오류, double로 캐스팅

+++ src/main/java/view/UiManager.java
@@ -31,7 +31,7 @@ public class UiManager {
     }
 
     public void printWinningStatistics(List<Integer> winningmatcher, int sumPrize) {
-        ResultView.printStatistics(winningmatcher, sumPrize / investAmount);
+        ResultView.printStatistics(winningmatcher, sumPrize / (double)investAmount);
     }
 
     private int takeCount(int investAmount) {
